updates to bring into line with Amy's shapetool branch
esp. redid mouseExit, mouseButtonDown, ect...
Now each action takes 'is clicked' value, some of this is redundant with existing functionality so at some point Amy and I will want to go through and sort out what is redundant and change.
the shapes are still very buggy and need to be sorted out